+++ tests/Jroc.Tests/String/JavaScript/String_PlusEquals_PlainAssignmentKeepsAssignedType.js
@@ -1,0 +1,6 @@
+"use strict";
+
+let s = 'a';
+s += 'b';
+s = 3;
+console.log(s, typeof s);